add seperate component for drag-drop and select from list in sample solvents
